# services/api-gateway/types/types.gen.go
// Package types provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package types

import (
	"time"
)

// Defines values for JobStatus.
const (
	JobStatusDead      JobStatus = "dead"
	JobStatusFailed    JobStatus = "failed"
	JobStatusPending   JobStatus = "pending"
	JobStatusQueued    JobStatus = "queued"
	JobStatusRejected  JobStatus = "rejected"
	JobStatusRunning   JobStatus = "running"
	JobStatusSucceeded JobStatus = "succeeded"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message *string `json:"message,omitempty"`
}

// Job defines model for Job.
type Job struct {
	CompletedAt *time.Time              `json:"completed_at"`
	CreatedAt   time.Time               `json:"created_at"`
	Id          string                  `json:"id"`
	Payload     *map[string]interface{} `json:"payload,omitempty"`
	Result      *map[string]interface{} `json:"result,omitempty"`
	Status      JobStatus               `json:"status"`
	Worker      string                  `json:"worker"`
}

// JobStatus defines model for JobStatus.
type JobStatus string

// Priority defines model for Priority.
type Priority = float32

// SubmitJobRequest defines model for SubmitJobRequest.
type SubmitJobRequest struct {
	Autorun            *bool                  `json:"autorun,omitempty"`
	Entrypoint         *[]string              `json:"entrypoint,omitempty"`
	IdleTimeoutSeconds *float32               `json:"idle_timeout_seconds,omitempty"`
	Image              string                 `json:"image"`
	MaxAttempts        *float32               `json:"max_attempts,omitempty"`
	Namespace          *string                `json:"namespace,omitempty"`
	Payload            map[string]interface{} `json:"payload"`
	Priority           *Priority              `json:"priority,omitempty"`
	TimeoutSeconds     *float32               `json:"timeout_seconds,omitempty"`
}

// SubmitJobResponse defines model for SubmitJobResponse.
type SubmitJobResponse struct {
	JobId  string    `json:"job_id"`
	Status JobStatus `json:"status"`
}

// SubmitJobJSONRequestBody defines body for SubmitJob for application/json ContentType.
type SubmitJobJSONRequestBody = SubmitJobRequest